requirements.txt
streamlit==1.32.0
psx-data-reader-master/
pytz==2024.1
beautifulsoup4==4.12.3
google-generativeai==0.3.2
requests==2.31.0
PyPDF2==3.0.1
pandas==2.2.1 